tools/gen_reference/requirements.txt
python-docx==0.8.11
python-pptx==0.6.21
